requirements.txt
requests>=2.31.0
structlog>=22.1.0

google-auth>=2.3.2

fastapi>=0.103.1
uvicorn>=0.22.0
gunicorn>=21.2.0
dnspython>=2.4.2
google-api-python-client>=2.101.0
google-cloud-speech>=2.21.0
google-cloud-secret-manager>=2.16.4
mysql-connector-python>=8.2.0

pydantic>=2.4.2

openai>=0.25.0
